miniscript-wip

+++ ext/cpp-spline/src/BSpline.cpp
@@ -24,13 +24,13 @@ void BSpline::_on_way_point_added()
 
 	for(int i=0; i<=_steps; i++)
 	{
-		double u=(double)i / (double)_steps;
+		float u=(float)i / (float)_steps;
 
 		add_node(interpolate(u, _way_points[pt], _way_points[pt+1], _way_points[pt+2], _way_points[pt+3]));
 	}
 }
 
-Vector BSpline::interpolate(double u, const Vector& P0, const Vector& P1, const Vector& P2, const Vector& P3)
+Vector BSpline::interpolate(float u, const Vector& P0, const Vector& P1, const Vector& P2, const Vector& P3)
 {
 	Vector point;
 	point=u*u*u*((-1) * P0 + 3 * P1 - 3 * P2 + P3) / 6;
