+++ ext/cpp-spline/src/Bezier.h
@@ -13,7 +13,7 @@ protected:
 	virtual void _on_way_point_added();
 
 protected:
-	Vector interpolate(double u, const Vector& P0, const Vector& P1, const Vector& P2, const Vector& P3);
+	Vector interpolate(float u, const Vector& P0, const Vector& P1, const Vector& P2, const Vector& P3);
 };
 
 #endif
