+++ ext/cpp-spline/src/Curve.h
@@ -23,14 +23,14 @@ protected:
 
 protected:
 	std::vector<Vector> _nodes;
-	std::vector<double> _distances;
+	std::vector<float> _distances;
 public:
 	Vector node(int i) const { return _nodes[i]; }
-	double length_from_starting_point(int i) const { return _distances[i]; }
+	float length_from_starting_point(int i) const { return _distances[i]; }
 	bool has_next_node(int i) const { return static_cast<int>(_nodes.size()) > i; }
 	int node_count() const {  return static_cast<int>(_nodes.size()); }
 	bool is_empty() { return _nodes.empty(); }
-	double total_length() const
+	float total_length() const
 	{
 		assert(!_distances.empty());
 		return _distances[_distances.size() - 1];
